requirements.txt
# Telegram Bot
python-telegram-bot[webhooks]==21.*

# LLM
anthropic>=0.42.0

# Async HTTP
aiohttp>=3.11.0
aiofiles>=24.1.0

# Database
aiosqlite>=0.20.0

# Voice Processing
ffmpeg-python>=0.2.0

# Trend Analysis
pytrends>=4.9.0
matplotlib>=3.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
